+++ src/cpx_io/cpx_system/cpx_base.py
@@ -2,8 +2,12 @@
 # TODO: Add Docstring
 '''
 import logging
+import struct
 
 from pymodbus.client import ModbusTcpClient
+from pymodbus.payload import BinaryPayloadDecoder
+from pymodbus.constants import Endian
+
 
 class CpxInitError(Exception):
     '''Error should be raised if a cpx-... module is instanciated without connecting it to a base module.
@@ -142,31 +146,51 @@ class CpxBase:
                 raise CpxInitError()
             return func(self, *args, **kwargs)
         return wrapper
-    
-    @staticmethod
-    def int_to_signed16(value: int):
-        '''Converts a int to 16 bit register where msb is the sign
-        with checking the range
-        '''
-        if (value <= -2**15) or (value > 2**15):
-            raise ValueError(f"Integer value {value} must be in range -32768...32767 (15 bit)")
         
-        if value >=0:
-            return value
+    @staticmethod
+    def _swap_bytes(registers):
+        swapped = []
+        for r in registers:
+            k = struct.pack('<H', r)
+            k = int.from_bytes(k, byteorder='big', signed=False)
+            swapped.append(k)
+        return swapped
+
+    @staticmethod
+    def _decode_string(registers):
+        # _swap_bytes has to be used because of a bug in pymodbus! 
+        # Byteorder does not work for strings. https://github.com/riptideio/pymodbus/issues/508
+        decoder = BinaryPayloadDecoder.fromRegisters(CpxBase._swap_bytes(registers), byteorder=Endian.BIG) 
+        return decoder.decode_string(34).decode('ascii').strip("\x00")
+
+    @staticmethod
+    def _decode_int(registers, type='uint16'):
+        decoder = BinaryPayloadDecoder.fromRegisters(registers[::-1], byteorder=Endian.BIG)
+        if type == "uint8":
+            return decoder.decode_8bit_uint()
+        elif type == "uint16":
+            return decoder.decode_16bit_uint()
+        elif type == "uint32":
+            return decoder.decode_32bit_uint()
+        if type == "int8":
+            return decoder.decode_8bit_int()
+        elif type == "int16":
+            return decoder.decode_16bit_int()
+        elif type == "int32":
+            return decoder.decode_32bit_int()
+        elif type == "bool":
+            return bool(decoder.decode_bits(0))
         else:
-            return 2**15 | ((value - 2**16) & ((2**16 - 1) // 2))
-    
-    @staticmethod
-    def signed16_to_int(value: int):
-        '''Converts a 16 bit register where msb is the sign to python signed int
-        by computing the two's complement 
-        '''
-        if value > 0xFFFF:
-            raise ValueError(f"Value {value} must not be bigger than 16 bit")
+            raise NotImplementedError(f"Type {type} not implemented")
         
-        if (value & (2**15)) != 0:        # if sign bit is set
-            value = value - 2**16       # compute negative value
-        return value
+    @staticmethod
+    def _decode_hex(registers, type='uint16'):
+        decoder = BinaryPayloadDecoder.fromRegisters(registers, byteorder=Endian.BIG)
+        if type == "uint16":
+            return format(decoder.decode_16bit_uint(), "#010x")
+        else:
+            raise NotImplementedError(f"Type {type} not implemented")
+        
     
     def __del__(self):
         self.client.close()